raise on duplicate key (GSI-657)
switched to ruamel.yaml for yaml parsing
