+++ .pre-commit-config.yaml
@@ -28,6 +28,7 @@ repos:
       - id: end-of-file-fixer
         exclude: '.*\.json|example_config.yaml'
       - id: check-yaml
+        exclude: DecodeError.duplicate_resource_id.datapack.yaml
       - id: check-added-large-files
       - id: check-ast
       - id: check-json
