+++ examples/datapack/invalid/simple_relations/DecodeError.duplicate_resource_id.datapack.yaml
@@ -1,0 +1,23 @@
+# Using a duplicate resource id:
+datapack: 0.2.0
+resources:
+  File:
+    example_file_a:
+      content:
+        filename: example_file_a.fastq
+        format: FASTQ
+        checksum: 1a5ac10ab42911dc0224172c118a326d9a4c03969112a2f3eb1ad971e96e92b8
+        size: 12321
+    example_file_a: # <-
+      content:
+        filename: example_file_a.fastq
+        format: FASTQ
+        checksum: 1a5ac10ab42911dc0224172c118a326d9a4c03969112a2f3eb1ad971e96e92b8
+        size: 12321
+  Dataset:
+    example_dataset_1:
+      content:
+        dac_contact: dac@example.org
+      relations:
+        files:
+          - example_file_a
