+++ pyproject.toml
@@ -27,7 +27,7 @@ description = "Make your JSON Schemas sociable and create linked data model."
 dependencies = [
     "pydantic >=2, <3",
     "pydantic_settings >=2, <3",
-    "PyYAML >=6.0.1, <7",
+    "ruamel.yaml ~= 0.18.6",
     "jsonschema >=4.19.2, <5",
     "immutabledict >=3.0.0, <4",
 ]
