+++ src/schemapack/exceptions.py
@@ -18,13 +18,26 @@
 
 from abc import ABC
 from dataclasses import dataclass
-from typing import Optional
+from pathlib import Path
+from typing import Literal, Optional
 
 import pydantic_core
 
 
 class BaseError(ABC, Exception):
     """Base class for all schemapack errors."""
+
+
+class DecodeError(BaseError, ValueError):
+    """Raised when decoding JSON or YAML data fails."""
+
+    def __init__(self, path: Path, assumed_format: Literal["JSON", "YAML"]):
+        super().__init__(
+            f"The file at '{path}' could not be decoded assuming {assumed_format}"
+            + " format."
+        )
+        self.path = path
+        self.assumed_format = assumed_format
 
 
 class SpecError(BaseError, ValueError):
