+++ src/schemapack/utils.py
@@ -22,28 +22,20 @@ import typing
 from contextlib import contextmanager
 from functools import lru_cache
 from pathlib import Path
-from typing import Any, Literal, TypeVar
+from typing import Any, TypeVar
 
 import jsonschema
 import jsonschema.exceptions
 import jsonschema.protocols
 import jsonschema.validators
-import yaml
+import ruamel.yaml
 from immutabledict import immutabledict
 from pydantic import GetCoreSchemaHandler
 from pydantic_core import core_schema
 
+from schemapack.exceptions import DecodeError
 
-class DecodeError(ValueError):
-    """Raised when decoding JSON or YAML data fails."""
-
-    def __init__(self, path: Path, assumed_format: Literal["JSON", "YAML"]):
-        super().__init__(
-            f"The file at '{path}' could not be decoded assuming {assumed_format}"
-            + " format."
-        )
-        self.path = path
-        self.assumed_format = assumed_format
+yaml = ruamel.yaml.YAML(typ="safe")
 
 
 def read_json_or_yaml(path: Path) -> dict:
@@ -66,8 +58,8 @@ def read_json_or_yaml(path: Path) -> dict:
         else:
             # Even if the file ending does not indicate YAML, we try to parse it as YAML
             try:
-                data = yaml.safe_load(file)
-            except yaml.YAMLError as error:
+                data = yaml.load(file)
+            except ruamel.yaml.YAMLError as error:
                 raise DecodeError(
                     path=path,
                     assumed_format="YAML",
