+++ tests/test_main.py
@@ -20,7 +20,12 @@ from pathlib import Path
 
 import pytest
 
-from schemapack.exceptions import BaseError, DataPackSpecError, ValidationError
+from schemapack.exceptions import (
+    BaseError,
+    DataPackSpecError,
+    DecodeError,
+    ValidationError,
+)
 from schemapack.main import load_and_validate
 from tests.fixtures.examples import (
     INVALID_DATAPACK_PATHS,
@@ -54,6 +59,8 @@ def test_load_and_validate_invalid(name: str, path: Path):
     # Either a DataPackSpecError or ValidationError is expected:
     if error_type == "DataPackSpecError":
         assert isinstance(exception_info.value, DataPackSpecError)
+    elif error_type == "DecodeError":
+        assert isinstance(exception_info.value, DecodeError)
     else:
         assert isinstance(exception_info.value, ValidationError)
         error_records = exception_info.value.records
